+++ webhooks/change/terraform-aws-okta-profile-webhook/src/index.ts
@@ -1,9 +1,10 @@
 import { APIGatewayProxyHandler } from 'aws-lambda'
 import { verify } from '@indent/webhook'
-import * as Indent from '@indent/types'
+import { Event } from '@indent/types'
 import axios from 'axios'
 
 import { getToken } from './utils/okta-auth'
+import * as oktaProfiles from './capabilities/okta-profiles'
 
 export const handle: APIGatewayProxyHandler = async function handle(event) {
   try {
@@ -28,7 +29,7 @@ export const handle: APIGatewayProxyHandler = async function handle(event) {
   console.log(JSON.stringify(events, null, 2))
 
   await Promise.all(
-    events.map((auditEvent: Indent.Event) => {
+    events.map((auditEvent: Event) => {
       let { actor, event, resources } = auditEvent
 
       console.log(
@@ -54,82 +55,4 @@ export const handle: APIGatewayProxyHandler = async function handle(event) {
     statusCode: 200,
     body: '{}',
   }
-}
-
-const OKTA_DOMAIN = process.env.OKTA_DOMAIN
-
-async function addUserToGroup({ user, group }) {
-  const { Authorization } = await getToken()
-  return await axios({
-    method: 'put',
-    url: `https://${OKTA_DOMAIN}/api/v1/groups/${group}/users/${user}`,
-    headers: {
-      Authorization,
-      Accept: 'application/json',
-      'Content-Type': 'application/json',
-    },
-  })
-}
-
-async function removeUserFromGroup({ user, group }) {
-  const { Authorization } = await getToken()
-  return await axios({
-    method: 'delete',
-    url: `https://${OKTA_DOMAIN}/api/v1/groups/${group}/users/${user}`,
-    headers: {
-      Authorization,
-      Accept: 'application/json',
-      'Content-Type': 'application/json',
-    },
-  })
-}
-
-const okta = { addUserToGroup, removeUserFromGroup }
-
-async function grantPermission(auditEvent: Indent.Event) {
-  const { event, actor, resources } = auditEvent
-  const user = getOktaIdFromResources(resources, 'user')
-  const group =
-    getOktaIdFromResources(resources, 'app') ||
-    getOktaIdFromResources(resources, 'group')
-
-  let result = await okta.addUserToGroup({ user, group })
-  console.log({
-    event,
-    actor,
-    resources,
-    success: result.status >= 200 && result.status < 300,
-  })
-}
-
-async function revokePermission(auditEvent: Indent.Event) {
-  const { event, actor, resources } = auditEvent
-  const user = getOktaIdFromResources(resources, 'user')
-  const group =
-    getOktaIdFromResources(resources, 'app') ||
-    getOktaIdFromResources(resources, 'group')
-
-  let result = await okta.removeUserFromGroup({ user, group })
-
-  console.log({
-    event,
-    actor,
-    resources,
-    success: result.status >= 200 && result.status < 300,
-  })
-}
-
-function getOktaIdFromResources(
-  resources: Indent.Resource[],
-  kind: string
-): string {
-  return resources
-    .filter((r) => r.kind && r.kind.toLowerCase().includes(kind.toLowerCase()))
-    .map((r) => {
-      if (r.labels && r.labels.oktaId) {
-        return r.labels.oktaId
-      }
-
-      return r.id
-    })[0]
 }
